stml0xx: Update Sensorhub IOCTLs to match 8916
Some of the sensorhub IOCTL values on Surnia was adjusted to be compatible with
the 32-bit+64-bit hybrid architecture. These new values needs to be applied to Otus
as well for consistency. This is required for the sensorhub updater app to work
properly as it takes a pre-built native library which assumes one set of IOCTL
values only.

Change-Id: I5ba04c64d48087e703dc8adc47db7f3114f7db37
Signed-off-by: Jason Jae Hyun Koh <jason.koh@motorola.com>
Reviewed-on: http://gerrit.mot.com/698068
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Robert Wiegand <Robert.P.Wiegand@motorola.com>
Reviewed-by: Daniel Moore <dtmoore@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
SLTApproved: Iyengar Sriram Kaushik <sriramkaushik@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>